+++ Sources/JSONSchemaMacro/Schemable/SchemaGenerator.swift
@@ -38,7 +38,7 @@ struct EnumSchemaGenerator {
     }
 
     if let annotationArguments = attributes.arguments(for: "SchemaOptions") {
-      codeBlockItem = TypeSpecificOptions.apply(annotationArguments, to: codeBlockItem, for: "SchemaOptions")
+      codeBlockItem = SchemaOptionsGenerator.apply(annotationArguments, to: codeBlockItem, for: "SchemaOptions")
     }
 
     let variableDecl: DeclSyntax = """
@@ -111,11 +111,11 @@ struct SchemaGenerator {
       "JSONObject { \(CodeBlockItemListSyntax(statements, separator: .newline)) }"
 
     if let annotationArguments = attributes.arguments(for: "SchemaOptions") {
-      codeBlockItem = TypeSpecificOptions.apply(annotationArguments, to: codeBlockItem, for: "SchemaOptions")
+      codeBlockItem = SchemaOptionsGenerator.apply(annotationArguments, to: codeBlockItem, for: "SchemaOptions")
     }
 
     if let objectArguemnts = attributes.arguments(for: "ObjectOptions") {
-      codeBlockItem = TypeSpecificOptions.apply(objectArguemnts, to: codeBlockItem, for: "ObjectOptions")
+      codeBlockItem = SchemaOptionsGenerator.apply(objectArguemnts, to: codeBlockItem, for: "ObjectOptions")
     }
 
     let variableDecl: DeclSyntax = """
