+++ Sources/JSONSchemaMacro/Schemable/SchemableMember.swift
@@ -63,12 +63,12 @@ struct SchemableMember {
 
     // Apply schema options if present
     if let annotationArguments = annotationArguments {
-      codeBlock = TypeSpecificOptions.apply(annotationArguments, to: codeBlock, for: "SchemaOptions")
+      codeBlock = SchemaOptionsGenerator.apply(annotationArguments, to: codeBlock, for: "SchemaOptions")
     }
 
     // Apply type-specific options if present
     if let typeSpecificArguments = typeSpecificArguments {
-      codeBlock = TypeSpecificOptions.apply(typeSpecificArguments, to: codeBlock, for: type.description)
+      codeBlock = SchemaOptionsGenerator.apply(typeSpecificArguments, to: codeBlock, for: type.description)
     }
 
     var block: CodeBlockItemSyntax = """
